Abstract reusable methods from sources
